+++ frontend/app/components/profile/helper/index.tsx
@@ -1,3 +1,0 @@
-import ManageBalanceSheet from "./ManageBalanceSheet";
-
-export { ManageBalanceSheet };