+++ frontend/app/components/profile/index.tsx
@@ -1,5 +1,6 @@
 import AccountSummaryCard from "./AccountSummaryCard";
 import PortfolioOverviewCard from "./PortfolioOverviewCard";
 import PortfolioPieChart from "./PortfolioPieChart";
+import ManageBalanceSheet from "./ManageBalanceSheet";
 
-export { AccountSummaryCard, PortfolioOverviewCard, PortfolioPieChart };
+export { AccountSummaryCard, PortfolioOverviewCard, PortfolioPieChart, ManageBalanceSheet };