+++ frontend/app/components/stock/index.tsx
@@ -1,5 +1,8 @@
-import StockAddPosition from "@/app/components/stock/StockAddPosition";
-import StockChart from "@/app/components/stock/StockChart";
-import StockList from "@/app/components/stock/StockList";
+import StockAddPosition from "./StockAddPosition";
+import StockChart from "./StockChart";
+import StockList from "./StockList";
+import StockDesktopTable from "./StockDesktopTable";
+import StockMobileCard from "./StockMobileCard";
+import QuantitySlider from "./QuantitySlider";
 
-export { StockAddPosition, StockChart, StockList };
+export { StockAddPosition, StockChart, StockList, StockDesktopTable, StockMobileCard, QuantitySlider };