+++ frontend/app/lib/formatter.tsx
@@ -1,14 +1,14 @@
 const Formatter = {
-    toLocale(num: number | any): string {
+    toLocale(num: number | any, keepDisplaying: boolean = false): string {
         if (typeof num !== "number")
             return "-";
 
-        return num === 0 ? "-" : num.toLocaleString("id-ID", { maximumFractionDigits: 2, minimumFractionDigits: 0 });
+        return (num === 0 && !keepDisplaying) ? "-" : num.toLocaleString("id-ID", { maximumFractionDigits: 2, minimumFractionDigits: 0 });
     },
 
     toCurrency(num: number | any): string {
         if (typeof num !== "number")
-            return "-";
+            return "Rp0";
 
         return num === 0 ? "-" : num.toLocaleString("id-ID", { style: "currency", currency: "IDR", maximumFractionDigits: 2, minimumFractionDigits: 0 });
     },